cancel two days prior profile

+++ ReservationDetails.cs
@@ -145,30 +145,67 @@ namespace GroupProjectDJT
 
 		}
 
-
-		
-
-
-
-
-
-private void cancelReservationButton_Click(object sender, EventArgs e)
+        private void cancelReservationButton_Click(object sender, EventArgs e)
 		{
-			var confirmResult = MessageBox.Show(
-				"Are you sure you want to cancel this reservation?",
-			" This cannot be undone!",
-				MessageBoxButtons.YesNo
-			);
+            if (canCancel(ReservationId))
+            {
+                var confirmResult = MessageBox.Show(
+                    "Are you sure you want to cancel this reservation?",
+                    " This cannot be undone!",
+                    MessageBoxButtons.YesNo
+                );
 
-			if (confirmResult == DialogResult.Yes)
-			{
-				deleteReservation();
-			}
-			else
-			{
-				//do nothing
-			}
-		}
+                if (confirmResult == DialogResult.Yes)
+                {
+                    deleteReservation();
+                }
+                else
+                {
+                    //do nothing
+                }
+            }
+            else
+            {
+                MessageBox.Show("Sorry, you can only cancel a reservation two days prior.");
+            }
+
+        }
+
+        private bool canCancel(int reservationId)
+        {
+            string connStr = "server=157.89.28.130;user=ChangK;database=csc340;port=3306;password=Wallace#409;";
+            MySql.Data.MySqlClient.MySqlConnection conn3 = new MySql.Data.MySqlClient.MySqlConnection(connStr);
+
+
+            Console.WriteLine("Connecting to MySQL...");
+            conn3.Open();
+            string sql = $@"SELECT
+	e.date
+FROM
+	djt_reservation r
+INNER JOIN
+	djt_event e
+ON
+	r.eventID = e.eventID
+WHERE
+	r.id = @reservationId;";
+            MySql.Data.MySqlClient.MySqlCommand cmd = new MySql.Data.MySqlClient.MySqlCommand(sql, conn3);
+
+            cmd.Parameters.AddWithValue("@reservationId", reservationId.ToString());
+
+            MySqlDataReader myReader = cmd.ExecuteReader();
+
+            myReader.Read();
+
+            DateTime dt = DateTime.Parse(myReader["date"].ToString());
+            DateTime today = DateTime.Now;
+
+
+            myReader.Close();
+            conn3.Close();
+
+            return (dt - today).Days > 2;
+        }
 
 		private void deleteReservation()
 		{
